# apidoc.yaml
swagger: "2.0"
info:
  description: "Bookstore RESTfull APIs"
  version: "1.0.0"
  title: "Ticket Management APIs"
  contact:
    email: "kittaporn.jp@gmail.com"
host: "localhost:3000"
basePath: "/"
...
schemes:
- "http"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "This will accept JWT token that came from login api"
paths:
  /users:
    get:
      tags:
      - "Users"
      summary: "Get logged in user information"
      description: "This api will return you a bunch of user personal information and user's orders"
      operationId: "getUserInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
        - Bearer: []
      responses:
        200:
          description: The returned data of logged in user
          schema: 
            $ref: '#/definitions/UserInformation'
    post:
      tags:
      - "Users"
      summary: "Create new user"
      description: "This api will recieve user information and store it into database"
      operationId: "CreateUser"
      parameters:
      - in: "body"
        name: "body"
        description: "The required parameters to create a new ticket as following"
        required: true
        schema:
          $ref: "#/definitions/CreateUserData"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: OK
    delete:
      tags:
      - "Users"
      summary: "Delete user"
      description: "This api will delete current logged in user"
      operationId: "SeleteUser"
      security:
        - Bearer: []
      parameters:
      - in: "body"
        name: "body"
        description: "The required parameters to delete current user as following"
        required: true
        schema:
          $ref: "#/definitions/CreateUserData"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: OK
...
definitions:
...
  CreateUserData:
    type: "object"
    properties:
      username:
        type: "string"
        example: "MarkWahlberg"
      password:
        type: "string"
        example: "1212312121"
      name:
        type: "string"
        example: "Mark"
      surname:
        type: "string"
        example: "Wahlberg"
      date_of_birth:
        type: "string"
        example: "1996-09-23"
  UserInformation:
    type: "object"
    properties:
      name:
        type: "string"
        example: "John"
      surname:
        type: "string"
        example: "Wahlberg"
      date_of_birth:
        type: "string"
        example: "1996-09-23"
      books:
        type: "array"
        items: 
          type: integer
        example:
          - 1
          - 4